email_mapping/pyproject.toml
[tool.poetry]
name = "odl_etl.email_mapping"
version = "0.1.0"
description = ""
authors = ["Tobias Macey <tmacey@mit.edu>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.6.1"
psycopg2-binary = "^2.8.6"
pyyaml = "^5.3.1"
pyarrow = "^1.0.1"
pandas = "^1.1.3"
logbook = "^1.5.3"
s3fs = "^0.5.1"
boto3 = "*"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "*"
pyformat = "*"
isort = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
